Remove 31 example and update 17 example
